Use freeipa_domain on all the systems, not just on client
While reinstalling the freeipa server of gluster, I was surprised
to see it was using rax.gluster.org. It turn out that it
was installed as "freeipa.gluster.org", and then later got renamed
to freeipa.rax.gluster.org, and so I didn't see that ansible would
auto detect the domain incorrectly.